services/ai-engine/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
torch
transformers
python-multipart
Pillow
openai-whisper